feat: test file names refactored
